Fix styling for RT portal iframe
The RT-Portal iframe in the Admin page was being squashed and not laid out
properly due to browser style-sheets. Adding explicit style fixes the layout
to use the space on the page consistently with other components.

While fixing the styling of the portal I found that the rt-portal classes in
the elevator-light theme admin css were unused, and wouldn't effect the contents
of the iframe due to iframe source not inheriting styles from parent document.

The unused styles were pulled in with the responsive styling from the
base-responsive branch in commit bb895ab211600e697ac8443b0a367ed0f2e75b0f but
haven't been used.